fix: StackBadge on Storybook
