+++ .storybook/preview.js
@@ -1,9 +1,4 @@
+import "../styles/globals.css";
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
 }
